Move ConfigProvider to subnamespace
Moves it to the `Zend\Expressive\Router\AuraRouter` subnamespace, to
prevent potential conflicts with other config providers under the
`Zend\Expressive\Router` namespace.
